Read dependency data used by buildDependencySetUsingLatest from IndexedDB
Reading the data from DB allows eventually dropping the reference to
whole list of ThunderstoreMods passed as an argument.

Unfortunately this requires bringing in a dependency to GameManager.
It's required currently to know which community's package list should
be used to look for the mods. Ideally the packages and listings would
be separated, but that's outside of the scope of the current task.
Additionally it would change the behaviour, as currently
inter-community dependencies aren't allowed.